filesystem(user)

+++ src/user/user.h
@@ -9,6 +9,6 @@ struct sysret {
 
 void putchar(char ch);
 int getchar(void);
-int readfile(const char *filename, char *buf, int len);
-int writefile(const char *filename, const char *buf, int len);
+int readfile(const char *filename, char *buf, uint64_t len);
+int writefile(const char *filename, const char *buf, uint64_t len);
 __attribute__((noreturn)) void exit(void);
